requirements.txt
asgiref==3.8.1
build==1.2.2.post1
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
cloudinary==1.44.1
colorama==0.4.6
Django==5.1.4
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
idna==3.10
packaging==25.0
pillow==11.2.1
pip-tools==7.4.1
PyJWT==2.9.0
pyproject_hooks==1.2.0
python-dotenv==1.1.0
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
stripe==12.2.0
typing_extensions==4.14.0
tzdata==2024.2
urllib3==2.5.0
whitenoise==6.9.0
